refactor: move response dataclass to core
