+++ tests/test_async_sc_connection.py
@@ -9,7 +9,8 @@ from sc_client.constants import common
 from sc_client.constants.common import RequestType, ScEventType
 from sc_client.constants.config import MAX_PAYLOAD_SIZE
 from sc_client.core.async_sc_connection import AsyncScConnection
-from sc_client.models import AsyncScEvent, Response
+from sc_client.core.response import Response
+from sc_client.models import AsyncScEvent
 from sc_client.sc_exceptions import ErrorNotes, PayloadMaxSizeError, ScServerError
 from sc_client.testing import ResponseCallback, WebsocketStub, websockets_client_connect_patch
